Moved game loop to backend.py
